feat: rename new `Explain` sheet to `Portal` (#4776)
* feat: non-hook to control Explain sheet

* feat: cover non-ds with ds components

* feat: rename Explain to Portal

* feat: rename methods

* feat: use extension, document

* fix: format
